+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [8.1.0] - 2026-03-09 
+
+### Changed
+
+- Replace circuar references in DOM with weak references to improve GC times https://github.com/Textualize/textual/pull/6410
+- When animating an attribute a second time, the original `on_complete` is now called https://github.com/Textualize/textual/pull/6410
+
+### Added
+
+- Added experimental `App.PAUSE_GC_ON_SCROLL_` boolean (disabled by default) https://github.com/Textualize/textual/pull/6410
+
 ## [8.0.2] - 2026-03-03
 
 ### Changed
@@ -3370,6 +3381,7 @@ https://textual.textualize.io/blog/2022/11/08/version-040/#version-040
 - New handler system for messages that doesn't require inheritance
 - Improved traceback handling
 
+[8.1.0]: https://github.com/Textualize/textual/compare/v8.0.2...v8.1.0
 [8.0.2]: https://github.com/Textualize/textual/compare/v8.0.1...v8.0.2
 [8.0.1]: https://github.com/Textualize/textual/compare/v8.0.0...v8.0.1
 [8.0.0]: https://github.com/Textualize/textual/compare/v7.5.0...v8.0.0